rename C++ headers with hpp extension
