+++ src/tool/analyser/main.cpp
@@ -4,10 +4,10 @@
 #include <vector>
 #include <ctime>
 
-#include <loudnessLibrary.h>
-#include <SoundFile.h>
-#include <WriteXml.h>
-#include <ProcessFile.h>
+#include <loudnessLibrary.hpp>
+#include <SoundFile.hpp>
+#include <WriteXml.hpp>
+#include <ProcessFile.hpp>
 
 bool showProgress = false;
 bool showResults = false;
